Adds points constraint to database
When creating the action to update all users, I realized that the
most efficient way to do it was to insert the random points using
the pg function to generate randon number.

Since this action will happen inside PG, it was more secure to add
the constraint not only in the changeset, but also at the DB using
migrations.