Creates the Genserver that updates user points and max points
The initial idea was to update the Users one by one, but since we
have a million users, the performance was too poor, so I did an
update all, passing a pg command that allowed the creation of the
random number directly on the database.

Updating user by user, the operation took almost 10 minutes. Using
the fragment, the operation took a little less than 3 seconds to
complete.

Another decision that I made on this commit was to pass the update
time as opts in the start link. That way, the testability was
improved.

About code organization, I know that usually people separate Client
and Server on different files in the genserver, but since this is a
small project, I decided to leave both on the same file.

In this commit I also intended to use the changeset to make the user
update, but after I realized that I won't use it anymore, I was
tempted to erase from the codebase, but I decided to leave since
using it is best practice.