Adds get_user function to GenServer
This function is pretty straightforward. It gets the first 2 users
that has points higher than the max_points, return those users along
with the last timestamp and update the timestamp.

I also added a fragile test, that depends on the randomnes of the
max number to work, but I leave a comment explaining how I would
fix this issue.