Adds web interface
In this commit I worked adding the web interface. The router, controller
and view.
The controller points at the genserver function `get_users()`.

Until the last commit all the functions at the genserver needed PID.
I had to do this to facilitate the tests, but when I needed to call
the `get_user` function I realized that I didn't had the pid to make
the function work, so I had to name the genserver and leave this
as a opts for the start link.